app/requirements.txt
# Database connection
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Data processing
pandas==2.1.3
numpy==1.26.2

# Data validation
pydantic==2.5.2
typing-extensions==4.8.0

# Date handling
python-dateutil==2.8.2
pytz==2023.3.post1

# Environment and config
python-dotenv==1.0.0

# Testing
pytest==7.4.3

# Logging and monitoring
loguru==0.7.2

# Production server
gunicorn==21.2.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0